Início do confirmar exclusão(artigo)

+++ src/controllers/excluir_artigo.php
@@ -1,22 +1,34 @@
 <?php
-    require_once '../assets/db/connect.php';
-    session_start();
+    $confirm = 0;
+    require_once "confirmar_exclusao.php";
 
-    if(isset($_GET['id_artigo'])){
+    if($confirm == 1){
+        require_once '../assets/db/connect.php';
+        session_start();
 
-        unlink($_GET['url_img']);
+        if(isset($_GET['id_artigo'])){
 
-        $result_evento = mysqli_query($conexao, "SELECT * FROM `artigos` WHERE `id` = '" . $_GET['id_artigo'] . "'");
-        $numero_result = mysqli_num_rows($result_evento);
+            unlink($_GET['url_img']);
 
-        if($numero_result != 0){
-            mysqli_query($conexao, "DELETE FROM `artigos` WHERE `id` =  '" . $_GET['id_artigo'] . "'");
-            ?>
-            <script>
-                alert("Artigo excluído!");
-                window.location.replace("../views/perfil.php");
-            </script>
-            <?php
+            $result_evento = mysqli_query($conexao, "SELECT * FROM `artigos` WHERE `id` = '" . $_GET['id_artigo'] . "'");
+            $numero_result = mysqli_num_rows($result_evento);
+
+            if($numero_result != 0){
+                mysqli_query($conexao, "DELETE FROM `artigos` WHERE `id` =  '" . $_GET['id_artigo'] . "'");
+                ?>
+                <script>
+                    alert("Artigo excluído!");
+                    window.location.replace("../views/perfil.php");
+                </script>
+                <?php
+            }else{
+                ?>
+                <script>
+                    alert("Este artigo não existe...");
+                    window.location.replace("../views/perfil.php");
+                </script>
+                <?php
+            }
         }else{
             ?>
             <script>
@@ -25,12 +37,5 @@
             </script>
             <?php
         }
-    }else{
-        ?>
-        <script>
-            alert("Este artigo não existe...");
-            window.location.replace("../views/perfil.php");
-        </script>
-        <?php
     }
 ?>
